+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Fixes `ProcessStore.fetch/1`. Now the function is able to retrieve the value from the process
+  dictionary when the process is nested into a process tree.
 
 ## [0.1.0] - 2019-11-26
 ### Added